Fixed thread stack allocation for Microblaze
The mb-gcc compiler seems to store certain values (e.g. the return address) at
locations above the current stack frame. This is not an issue when calling a
function from another function which has an appropriately prepared stack. When
creating a new thread, however, there is no prepared stack, and the first call
to the entry function causes the compiler to overwrite values which
immediately precede the memory location reserved by for the thread's stack.

This fix causes the stack pointer to point to a location 64 bytes beyond the
start of the threads stack frame; this way, there is enough free space for the
compiler to store its values on calling the entry function. This obviously
wastes memory for every thread; the offset value of 64 bytes could probably be
reduced after further analysis.