Added enable logic to xps_osif
The added logic performs the gating function for signals crossing from
the reconfigurable region to the static logic. This function was previously
handled by the bus macros, which are now deprecated by the new 12.x PR tool
chain. The signals are still named 'bmEnable' or similar because of their
'bus macro' origins.

The 'bmEnable' output of the xps_osif core remains operational for
compatibility reasons.