Add 12.3 ML605 refdesign with slower clocked ICAP.
This ml605_light_pr_slowicap is a workaround for the ml605_light_pr design not
meeting timing. The _pr_slowicap design's ICAP is clocked at 75 MHz instead of
the specified maximum of 100 MHz. This may impact reconfiguration times, though
testing is needed to confirm that.

An alternative (and better) solution would be to constrain the _pr reference
design's xhwicap core so that it can reach 100 Mhz. Any volunteers? ;-)