Split planahead build targets into multiple steps
There are three new build targets now:

complete (or 'planahead-all')
    This builds all bitstreams (initial and partial) for a ReconOS design
    where all threads are known in advance.

extensible-base (or 'planahead-base')
    Like 'complete', but builds only partial bitstreams for the first thread
    (in all slots). This creates something like an 'implemented reference
    design' for which partial bitstreams can be added later (see
    'extensible-threads').

extensible-threads (or 'planahead-threads')
    This target builds only the intital and partial bitstreams for the
    second and following DYNAMIC_THREADS, based on the reference
    implementation built in 'extensible-base'. This allows us to define and
    add threads (and create corresponding partial bitstreams) after the
    intital bitstream has been created (and deployed).

In other words, 'extensible-threads' promotes the static partition of the
design built in 'extensible-base', so that the partial bitstreams resulting
from 'extensible-threads' can be used with an FPGA configured with an intial
bitstream from 'extensible-base'.

This commit also involves a modularization of 'mkpatcl_partial.py'. Invoke
'mkpatcl_partial.py --help' for an overview.